requirements.txt
azure.functions
azure-functions-worker
scholarly
jsonpickle